pyproject.toml
[project]
name = "structflo-cser"
version = "0.1.0"
description = "YOLO-based chemical structure + label detector with synthetic data generation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chembl-webresource-client>=0.10.9",
    "decimer>=2.8.0",
    "flask>=3.1.3",
    "matplotlib>=3.10.8",
    "easyocr>=1.7",
    "numpy>=2.4.2",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.1.1",
    "pymupdf>=1.27.1",
    "rdkit>=2025.9.5",
    "huggingface_hub>=0.24.0",
    "packaging>=23.0",
    "requests>=2.31.0",
    "scipy>=1.17.0",
    "pandas>=2.0.0",
    "tqdm>=4.67.3",
    "ultralytics>=8.4.14",
]

[project.scripts]
sf-fetch-smiles         = "structflo.cser.data.smiles:main"
sf-generate             = "structflo.cser.generation.dataset:main"
sf-train                = "structflo.cser.training.trainer:main"
sf-detect               = "structflo.cser.inference.detector:main"
sf-viz                  = "structflo.cser.viz.labels:main"
sf-download-distractors = "structflo.cser.data.distractor_images:main"
sf-annotate             = "annotate.__main__:main"
sf-extract              = "structflo.cser.pipeline.cli:main"
sf-train-lps            = "structflo.cser.lps.train:main"
sf-eval-lps             = "structflo.cser.lps.evaluate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["structflo", "annotate"]

[dependency-groups]
dev = [
    "jupyterlab>=4.5.4",
    "pytest>=9.0.2",
    "ruff>=0.15.2",
]
